requirements.txt
aiofiles==23.2.1
aiogram==3.12.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
deepl==1.18.0
docx==0.2.4
frozenlist==1.4.1
idna==3.8
lxml==5.3.0
magic-filter==1.0.12
multidict==6.0.5
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
python-dotenv==1.0.1
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.11
